+++ applovin_max/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Versions
 
-## x.x.x
+## 4.0.1
 * Crash if plugin version is incompatible with native SDK version.
 * Add support for Yandex native ads.
 * Update listener APIs to allow for null values for unsetting.